AddHorizontalLine() and AddVerticalLine() #638 #639
