+++ src/ScottPlot/Plot/Plot.Obsolete.cs
@@ -86,6 +86,62 @@ namespace ScottPlot
             return scatter;
         }
 
+        [Obsolete("Use AddHorizontalLine() and customize the object it returns")]
+        public HLine PlotHLine(
+            double y,
+            Color? color = null,
+            double lineWidth = 1,
+            string label = null,
+            bool draggable = false,
+            double dragLimitLower = double.NegativeInfinity,
+            double dragLimitUpper = double.PositiveInfinity,
+            LineStyle lineStyle = LineStyle.Solid
+            )
+        {
+            var hline = new HLine()
+            {
+                position = y,
+                color = color ?? settings.GetNextColor(),
+                lineWidth = (float)lineWidth,
+                label = label,
+                DragEnabled = draggable,
+                lineStyle = lineStyle,
+                DragLimitMin = dragLimitLower,
+                DragLimitMax = dragLimitUpper
+            };
+            Add(hline);
+            return hline;
+        }
+
+        [Obsolete("Use AddHorizontalSpan() and customize the object it returns")]
+        public HSpan PlotHSpan(
+            double x1,
+            double x2,
+            Color? color = null,
+            double alpha = .5,
+            string label = null,
+            bool draggable = false,
+            bool dragFixedSize = false,
+            double dragLimitLower = double.NegativeInfinity,
+            double dragLimitUpper = double.PositiveInfinity
+            )
+        {
+            var axisSpan = new HSpan()
+            {
+                position1 = x1,
+                position2 = x2,
+                color = color ?? settings.GetNextColor(),
+                alpha = alpha,
+                label = label,
+                DragEnabled = draggable,
+                DragFixedSize = dragFixedSize
+            };
+            axisSpan.SetLimits(dragLimitLower, dragLimitUpper, null, null);
+
+            Add(axisSpan);
+            return axisSpan;
+        }
+
         [Obsolete("Use AddText() and customize the object it returns")]
         public Text PlotText(
             string text,
@@ -121,6 +177,61 @@ namespace ScottPlot
             };
             Add(plottableText);
             return plottableText;
+        }
+
+        [Obsolete("Use AddVerticalLine() and customize the object it returns")]
+        public VLine PlotVLine(
+            double x,
+            Color? color = null,
+            double lineWidth = 1,
+            string label = null,
+            bool draggable = false,
+            double dragLimitLower = double.NegativeInfinity,
+            double dragLimitUpper = double.PositiveInfinity,
+            LineStyle lineStyle = LineStyle.Solid
+            )
+        {
+            VLine axLine = new VLine()
+            {
+                position = x,
+                color = color ?? settings.GetNextColor(),
+                lineWidth = (float)lineWidth,
+                label = label,
+                DragEnabled = draggable,
+                lineStyle = lineStyle,
+                DragLimitMin = dragLimitLower,
+                DragLimitMax = dragLimitUpper
+            };
+            Add(axLine);
+            return axLine;
+        }
+
+        [Obsolete("Use AddVerticalSpan() and customize the object it returns")]
+        public VSpan PlotVSpan(
+            double y1,
+            double y2,
+            Color? color = null,
+            double alpha = .5,
+            string label = null,
+            bool draggable = false,
+            bool dragFixedSize = false,
+            double dragLimitLower = double.NegativeInfinity,
+            double dragLimitUpper = double.PositiveInfinity
+            )
+        {
+            var axisSpan = new VSpan()
+            {
+                position1 = y1,
+                position2 = y2,
+                color = color ?? settings.GetNextColor(),
+                alpha = alpha,
+                label = label,
+                DragEnabled = draggable,
+                DragFixedSize = dragFixedSize,
+            };
+            axisSpan.SetLimits(null, null, dragLimitLower, dragLimitUpper);
+            Add(axisSpan);
+            return axisSpan;
         }
     }
 }
